Added new file to compiler script and added a check if executable exists

+++ compile.py
@@ -1,7 +1,7 @@
 import os
 
 flags = ["-lSDL2", "-lSDL2_image"]
-files = ["Sprite.cpp", "SDL_Setup.cpp", "main.cpp", "Main.cpp", "Bullet.cpp", "Pattern.cpp", "Player.cpp"]
+files = ["Sprite.cpp", "SDL_Setup.cpp", "main.cpp", "Main.cpp", "Bullet.cpp", "Pattern.cpp", "Player.cpp", "GameStats.cpp"]
 
 command = "g++ "
 for file in files:
@@ -10,6 +10,8 @@ for file in files:
 for flag in flags:
 	command += flag + " "
 
-os.system("rm BulletHell")
+if(os.path.isfile('BulletHell')):
+	os.system("rm BulletHell")
+
 os.system(command[:-1])
 os.system("mv a.out BulletHell")
